+++ matrix-client/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0.0
+
+- Expose mkRequest and doRequest
+- Return DeviceID out of getTokenOwner
+
 ## 0.1.5.0
 
 - Convert userDisplayName into a Maybe value [#30](https://github.com/softwarefactory-project/matrix-client-haskell/issues/30)